ai-service/requirements.txt
flask
groq
PyPDF2
requests
python-dotenv
python-docx
python-pptx
